Registries reorg
